rewrite_agent_311/pyproject.toml
[project]
name = "rewrite_agent_exe"
version = "0.1.0"
description = "A Python project with browser automation and AI capabilities"
requires-python = ">=3.11"
authors = [
    {name = "Project Author", email = "guymorganb@gmail.com"}
]
dependencies = [
    "greenlet==3.1.1",
    "selenium",
    "webdriver-manager",
    "langchain-openai",
    "langchain-community",
    "python-dotenv",
    "ipython",
    "langchain==0.3.15",
    "openai",
    "langgraph",
    "langchain-core",
    "numpy==1.26.3",
    "opencv-python",
    "protobuf",
    "pyyaml",
    "cryptography",
    "psycopg2-binary",
    "lxml",
    "graphviz",
    "pillow",
    "python-magic",
    "ffmpeg-python",
    "tavily-python",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["rewrite_agent_exe"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I"]

[tool.ruff.isort]
known-first-party = ["rewrite_agent_exe"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["rewrite_agent_exe"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
